requirements.txt
asgiref==3.2.7
bootstrap4==0.1.0
config==0.5.0.post0
decouple==0.0.7
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.1
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
